Move CLI command registration into bounded contexts
Each bounded context now exports a register_commands() function
that wires its command groups onto the Click CLI. main.py becomes
a thin composition root that discovers and registers them.

Entry point renamed from 'consultamatron' to 'practice'.
